+++ app.py
@@ -1,9 +1,9 @@
 from flask import Flask, jsonify
-from muntrunk.parse import parse_entire_list
+from muntrunk.parse import parse_w2020
 
 app = Flask(__name__)
 
 
 @app.route("/")
 def get_entire_list():
-    return jsonify(parse_entire_list())
+    return jsonify(parse_w2020())
